# tests/contrato_openapi_v1.json
{"openapi": "3.1.0", "info": {"title": "FastAPI Arquetipo", "description": "Arquetipo Idiomatico Escalable de FastAPI", "version": "1.0.0"}, "paths": {"/api/autenticacion/registrar-cuenta": {"post": {"tags": ["Autenticacion"], "summary": "Registrar Cuenta", "operationId": "registrar_cuenta_api_autenticacion_registrar_cuenta_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/RegistroPeticion"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/autenticacion/iniciar-sesion": {"post": {"tags": ["Autenticacion"], "summary": "Iniciar Sesion", "operationId": "iniciar_sesion_api_autenticacion_iniciar_sesion_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/LoginPeticion"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/autenticacion/refrescar-token": {"post": {"tags": ["Autenticacion"], "summary": "Refrescar Token", "operationId": "refrescar_token_api_autenticacion_refrescar_token_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/RefrescarTokenPeticion"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/autenticacion/cerrar-sesion": {"post": {"tags": ["Autenticacion"], "summary": "Cerrar Sesion", "operationId": "cerrar_sesion_api_autenticacion_cerrar_sesion_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/RefrescarTokenPeticion"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/autenticacion/cerrar-sesion-todos": {"post": {"tags": ["Autenticacion"], "summary": "Cerrar Sesion Todos", "operationId": "cerrar_sesion_todos_api_autenticacion_cerrar_sesion_todos_post", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/autenticacion/solicitar-recuperacion-clave": {"post": {"tags": ["Autenticacion"], "summary": "Solicitar Recuperacion Clave", "operationId": "solicitar_recuperacion_clave_api_autenticacion_solicitar_recuperacion_clave_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/SolicitarRecuperacionClavePeticion"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/autenticacion/restablecer-clave": {"post": {"tags": ["Autenticacion"], "summary": "Restablecer Clave", "operationId": "restablecer_clave_api_autenticacion_restablecer_clave_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/RestablecerClavePeticion"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/usuario/perfil": {"get": {"tags": ["Usuario"], "summary": "Obtener Mi Perfil", "operationId": "obtener_mi_perfil_api_usuario_perfil_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/usuario/{usuario_id}/perfil": {"get": {"tags": ["Usuario"], "summary": "Obtener Perfil Por Usuario Id", "operationId": "obtener_perfil_por_usuario_id_api_usuario__usuario_id__perfil_get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "usuario_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Usuario Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/usuarios": {"post": {"tags": ["Usuarios Gestion"], "summary": "Crear Usuario Gestion", "operationId": "crear_usuario_gestion_api_usuarios_post", "security": [{"OAuth2PasswordBearer": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UsuarioGestionCrearPeticion"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "get": {"tags": ["Usuarios Gestion"], "summary": "Listar Usuarios Gestion", "operationId": "listar_usuarios_gestion_api_usuarios_get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "pagina", "in": "query", "required": false, "schema": {"type": "integer", "default": 1, "title": "Pagina"}}, {"name": "tamano", "in": "query", "required": false, "schema": {"type": "integer", "default": 10, "title": "Tamano"}}, {"name": "correo", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Correo"}}, {"name": "rol", "in": "query", "required": false, "schema": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Rol"}}, {"name": "estado", "in": "query", "required": false, "schema": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Estado"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/usuarios/{usuario_id}": {"get": {"tags": ["Usuarios Gestion"], "summary": "Obtener Usuario Gestion", "operationId": "obtener_usuario_gestion_api_usuarios__usuario_id__get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "usuario_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Usuario Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["Usuarios Gestion"], "summary": "Actualizar Usuario Gestion", "operationId": "actualizar_usuario_gestion_api_usuarios__usuario_id__put", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "usuario_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Usuario Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UsuarioGestionActualizarPeticion"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["Usuarios Gestion"], "summary": "Eliminar Logico Usuario Gestion", "operationId": "eliminar_logico_usuario_gestion_api_usuarios__usuario_id__delete", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "usuario_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Usuario Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/usuarios/{usuario_id}/estado": {"patch": {"tags": ["Usuarios Gestion"], "summary": "Cambiar Estado Usuario Gestion", "operationId": "cambiar_estado_usuario_gestion_api_usuarios__usuario_id__estado_patch", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "usuario_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Usuario Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UsuarioGestionEstadoPeticion"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/estado": {"get": {"tags": ["Estado"], "summary": "Verificar Estado", "operationId": "verificar_estado_api_estado_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "LoginPeticion": {"properties": {"correo": {"type": "string", "title": "Correo"}, "clave": {"type": "string", "title": "Clave"}}, "type": "object", "required": ["correo", "clave"], "title": "LoginPeticion"}, "RefrescarTokenPeticion": {"properties": {"token_refresco": {"type": "string", "minLength": 20, "title": "Token Refresco"}}, "type": "object", "required": ["token_refresco"], "title": "RefrescarTokenPeticion"}, "RegistroPeticion": {"properties": {"nombre": {"type": "string", "maxLength": 100, "title": "Nombre"}, "apellidos": {"type": "string", "maxLength": 100, "title": "Apellidos"}, "correo": {"type": "string", "format": "email", "title": "Correo"}, "clave": {"type": "string", "maxLength": 255, "minLength": 6, "title": "Clave"}, "telefono": {"anyOf": [{"type": "string", "maxLength": 12}, {"type": "null"}], "title": "Telefono"}, "direccion": {"anyOf": [{"type": "string", "maxLength": 255}, {"type": "null"}], "title": "Direccion"}, "referencia": {"anyOf": [{"type": "string", "maxLength": 255}, {"type": "null"}], "title": "Referencia"}, "codigo_postal": {"anyOf": [{"type": "string", "maxLength": 10}, {"type": "null"}], "title": "Codigo Postal"}, "distrito_id": {"anyOf": [{"type": "string", "maxLength": 6}, {"type": "null"}], "title": "Distrito Id"}}, "type": "object", "required": ["nombre", "apellidos", "correo", "clave"], "title": "RegistroPeticion"}, "RestablecerClavePeticion": {"properties": {"token_recuperacion": {"type": "string", "minLength": 20, "title": "Token Recuperacion"}, "nueva_clave": {"type": "string", "maxLength": 255, "minLength": 6, "title": "Nueva Clave"}}, "type": "object", "required": ["token_recuperacion", "nueva_clave"], "title": "RestablecerClavePeticion"}, "SolicitarRecuperacionClavePeticion": {"properties": {"correo": {"type": "string", "title": "Correo"}}, "type": "object", "required": ["correo"], "title": "SolicitarRecuperacionClavePeticion"}, "UsuarioGestionActualizarPeticion": {"properties": {"nombre": {"type": "string", "maxLength": 100, "minLength": 1, "title": "Nombre"}, "apellidos": {"type": "string", "maxLength": 100, "minLength": 1, "title": "Apellidos"}, "correo": {"type": "string", "format": "email", "title": "Correo"}, "telefono": {"anyOf": [{"type": "string", "maxLength": 12}, {"type": "null"}], "title": "Telefono"}, "foto_url": {"anyOf": [{"type": "string", "maxLength": 500}, {"type": "null"}], "title": "Foto Url"}, "rol_id": {"type": "integer", "title": "Rol Id"}, "estado": {"type": "integer", "title": "Estado", "default": 1}}, "type": "object", "required": ["nombre", "apellidos", "correo", "rol_id"], "title": "UsuarioGestionActualizarPeticion"}, "UsuarioGestionCrearPeticion": {"properties": {"nombre": {"type": "string", "maxLength": 100, "minLength": 1, "title": "Nombre"}, "apellidos": {"type": "string", "maxLength": 100, "minLength": 1, "title": "Apellidos"}, "correo": {"type": "string", "format": "email", "title": "Correo"}, "telefono": {"anyOf": [{"type": "string", "maxLength": 12}, {"type": "null"}], "title": "Telefono"}, "foto_url": {"anyOf": [{"type": "string", "maxLength": 500}, {"type": "null"}], "title": "Foto Url"}, "rol_id": {"type": "integer", "title": "Rol Id"}, "estado": {"type": "integer", "title": "Estado", "default": 1}}, "type": "object", "required": ["nombre", "apellidos", "correo", "rol_id"], "title": "UsuarioGestionCrearPeticion"}, "UsuarioGestionEstadoPeticion": {"properties": {"estado": {"type": "integer", "title": "Estado"}}, "type": "object", "required": ["estado"], "title": "UsuarioGestionEstadoPeticion"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}, "securitySchemes": {"OAuth2PasswordBearer": {"type": "oauth2", "flows": {"password": {"scopes": {}, "tokenUrl": "api/autenticacion/iniciar-sesion"}}}}}}